pyproject.toml
[tool.poetry]
name = "diploma2"
version = "0.1.0"
description = ""
authors = ["Stepan Filippov <stef-2020@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
transformers = "^4.39.3"
torch = "^2.2.2"
aiogram = "^3.6.0"
python-dotenv = "^1.0.1"
googletrans = "4.0.0rc1"
emoji = "^2.12.1"
pymemcache = "^4.0.0"
deepl = "^1.18.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
